--- backend/business-logic/reset.js
@@ -1,8 +1,0 @@
-
-const Token = require('../data-access/db');
-
-const resetPasswordManager = {
-    
-        }
-
-module.exports = resetPasswordManager;
